requirements.txt
alembic==1.0.0
aniso8601==3.0.2
certifi==2018.8.24
chardet==3.0.4
Click==7.0
feedparser==5.2.1
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gcloud==0.17.0
googleapis-common-protos==1.5.3
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jws==0.1.3
Mako==1.0.7
MarkupSafe==1.0
oauth2client==3.0.0
protobuf==3.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycryptodome==3.4.3
PyJWT==1.6.4
Pyrebase==3.0.27
python-dateutil==2.7.3
python-editor==1.0.3
python-jwt==2.0.1
pytz==2018.5
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.11.0
SQLAlchemy==1.2.12
sqlalchemy-pagination==0.0.2
urllib3==1.23
Werkzeug==0.14.1
